+++ pkgs/google_generative_ai/lib/src/client.dart
@@ -17,6 +17,7 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 
+import 'error.dart';
 import 'version.dart';
 
 const clientName = 'genai-dart/$packageVersion';
@@ -61,6 +62,11 @@ final class HttpApiClient implements ApiClient {
       headers: await _headers(),
       body: _utf8Json.encode(body),
     );
+    if (response.statusCode >= 500) {
+      throw GenerativeAIException(
+          'Server Error [${response.statusCode}]: ${response.body}');
+    }
+
     return _utf8Json.decode(response.bodyBytes) as Map<String, Object?>;
   }
 
@@ -71,11 +77,7 @@ final class HttpApiClient implements ApiClient {
     final request = http.Request('POST', uri)
       ..bodyBytes = _utf8Json.encode(body)
       ..headers.addAll(await _headers());
-    // TODO: When updating min SDK remove workaround.
-    final httpClient = _httpClient;
-    final response = httpClient == null
-        ? await request.send()
-        : await httpClient.send(request);
+    final response = await (_httpClient?.send(request) ?? request.send());
     if (response.statusCode != 200) {
       final body = await response.stream.bytesToString();
       // Yeild a potential error object like a normal result for consistency