+++ samples/dart/bin/system_instructions.dart
@@ -12,35 +12,35 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:google_generative_ai/google_generative_ai.dart';
 
-Future<void> generate(GenerationConfig? generationConfig, String apiKey) async {
-  final model = GenerativeModel(
-      model: 'gemini-1.5-flash-latest',
-      apiKey: apiKey,
-      generationConfig: generationConfig);
-  final prompt = 'One, two, three, ';
-  print('Prompt: $prompt');
-  final content = [Content.text(prompt)];
-
-  print('Options: ${jsonEncode(generationConfig?.toJson())}');
-
-  final response = await model.generateContent(content);
-  print('Response:');
-  print(response.text);
-}
-
-Future<void> main() async {
+final apiKey = () {
   final apiKey = Platform.environment['GEMINI_API_KEY'];
   if (apiKey == null) {
     stderr.writeln(r'No $GEMINI_API_KEY environment variable');
     exit(1);
   }
-  await generate(null, apiKey);
-  await generate(GenerationConfig(maxOutputTokens: 3), apiKey);
-  await generate(GenerationConfig(stopSequences: ['seven']), apiKey);
-  await generate(GenerationConfig(temperature: 0), apiKey);
+  return apiKey;
+}();
+
+Future<void> systemInstructions() async {
+  // [START system_instructions]
+  // Make sure to include this import:
+  // import 'package:google_generative_ai/google_generative_ai.dart';
+  final model = GenerativeModel(
+    model: 'gemini-1.5-flash',
+    apiKey: apiKey,
+    systemInstruction: Content.system('You are a cat. Your name is Neko.'),
+  );
+  final prompt = 'Good morning! How are you?';
+
+  final response = await model.generateContent([Content.text(prompt)]);
+  print(response.text);
+  // [END system_instructions]
+}
+
+void main() async {
+  await systemInstructions();
 }